Add hasResponded check to Message
This adds the capability to check if a message has been responded to, by `finish` or `requeue`.
It also makes these operations idempotent and NOOP after the first attempt.

In addition to this, it also makes `touch` a NOOP when a message has already been responded to.

Callers can now unconditinally `finish` or `requeue` without being concerned about any acknowledgemnt that has happened before.
Or they can simply check if a message has been responded to or not.

For backwards-compatility, the interface method defaults to returning `false`.